Move ActivityHeader connect
