+++ ui/src/components/activity/ActivityHeader/index.js
@@ -1,2 +1,1 @@
-import { ActivityHeader } from './connect';
-export { ActivityHeader };
+export { ActivityHeader } from './ActivityHeader';
