+++ ui/src/components/index.js
@@ -1,1 +1,2 @@
 export { ActivitySearch } from './activity/ActivitySearch';
+export { ActivityHeader } from './activity/ActivityHeader';
