+++ ui/src/implementations/kapture-backend/components/index.js
@@ -1,2 +1,3 @@
 export { Header } from '../../../components/app/Header';
 export { ActivitySearch } from './ActivitySearchConnected';
+export { ActivityHeader } from './ActivityHeaderConnected';
